Lots of changes:
Protocol Display now has forms and method to save comments and goals,
Currently getting the 405 error when it tries to save either.

Changes to addAssignedProtocol to retrieve more data from the API and populate the clientProfile more.

super quick homework templates page, just really to be there for the buttons on the clientProfile and nothing else.

still working on getting the step info for the assignedProtocols so we can do things like the action items