Big Changes
Upload File setup and working
Add Coclient + AddDependent Included however they do not work as the api still doesn't allow those to be patched into the users it seems.
Client Profile has the the tabs to filter
active, completed and archieved Protocols however as there isn't a status id yet in the protocols this is commented out but it's there.
CoClient name now redirects the user to the coClient.