Completion Report
Currently not Hooked up anywhere, need to figure out how/if we're going to hook it up to the reports page.

http://localhost:8080/completionReport
you can find it here if you want to have a look.

Until we get the completion dates working it won't show anything, as it's set to ignore any protocol with a nul completed Date. I chose this instead of status because at the end of the day what this function cares about is the actual day it was completed on and not the status.