fixes to the client profile bug
Issue came from the workaround to fix the problem where primary contacts were erased when we updated the coclients of the household.
It worked under the assumption that all households would have primary contacts. When they didn't it threw an error, though the action of adding or removing a coclient would still be acted on.