Household Filtering logic
Ended up redoing the way we did things here
instead of loading the data for questions on each coclient with a separate controller decided it'd be easier and better to
create a new class and that combined user+ questionResponseList and throw all of them into the initial load.
This way we can do all the filtering in the Html.
This'll make it easier to implement answering questions and the parent categories when we get around to that.