UI Updates for Protocol View, User Profile View, Household View, and Header
Protocol View Added filter logic to make sure it was only showing questions and answers for the Assigned protocol and not every question assigned to user.

Got rid of the Unanswered tab in the categories section of the Questions Tabs as it was not needed anymore

Household View: All Client Names where being displayed under the Questions Section, Got rid of that.

Removed the Homeworks Template from the Header dropdown as those don't exist anymore.